pyproject.toml
[project]
name = "videoseal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=16.0.1",
    "einops>=0.8.1",
    "ffmpeg-python>=0.2.0",
    "omegaconf>=2.3.0",
    "onnxscript>=0.5.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pycocotools>=2.0.10",
    "pytorch-msssim>=1.0.0",
    "pywavelets>=1.8.0",
    "safetensors>=0.6.2",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "tensorboard>=2.20.0",
    "tensorrt==10.14.1.48.post1",
    "timm==0.9.16",
    "torch==2.3.1",
    "tqdm>=4.67.1",
    "cupy-cuda13x",
    "requests"
]
